SLF4J 2: add a service provider
