+++ pom.xml
@@ -75,7 +75,7 @@
     </distributionManagement>
 
     <properties>
-        <slf4j.version>1.7.36</slf4j.version>
+        <slf4j.version>2.0.1</slf4j.version>
         <junit.version>5.9.1</junit.version>
         <mockito.version>4.6.1</mockito.version>
         <assertj.version>3.23.1</assertj.version>
@@ -203,7 +203,7 @@
                                     <version>${slf4j.version}</version>
                                 </artifactItem>
                             </artifactItems>
-                            <includes>org/slf4j/impl/*.java</includes>
+                            <includes>org/slf4j/simple/*.java</includes>
                             <excludes>**/Static*,**/SimpleLoggerFactory*</excludes>
                             <outputDirectory>${project.build.directory}/generated-sources/annotations</outputDirectory>
                         </configuration>
