+++ src/main/java/org/simplify4u/slf4jmock/MockServiceProvider.java
@@ -1,0 +1,55 @@
+package org.simplify4u.slf4jmock;
+
+import org.slf4j.ILoggerFactory;
+import org.slf4j.IMarkerFactory;
+import org.slf4j.helpers.BasicMarkerFactory;
+import org.slf4j.simple.StaticLoggerBinder;
+import org.slf4j.simple.StaticMDCBinder;
+import org.slf4j.spi.MDCAdapter;
+import org.slf4j.spi.SLF4JServiceProvider;
+
+/**
+ * Responsible for binding the {@link LoggerMock}, the {@link BasicMarkerFactory} and the
+ * {@link MDCMock}. This is used by the SLF4J API.
+ */
+public class MockServiceProvider implements SLF4JServiceProvider {
+
+  /**
+   * Declare the version of the SLF4J API this implementation is compiled against. The value of this
+   * field is modified with each major release.
+   */
+  // to avoid constant folding by the compiler, this field must *not* be final
+  public static String REQUESTED_API_VERSION = "2.0.99"; // !final
+
+  private ILoggerFactory loggerFactory;
+  private IMarkerFactory markerFactory;
+  private MDCAdapter contextAdapter;
+
+  @Override
+  public ILoggerFactory getLoggerFactory() {
+    return loggerFactory;
+  }
+
+  @Override
+  public IMarkerFactory getMarkerFactory() {
+    return markerFactory;
+  }
+
+  @Override
+  public MDCAdapter getMDCAdapter() {
+    return contextAdapter;
+  }
+
+  @Override
+  public String getRequestedApiVersion() {
+    return REQUESTED_API_VERSION;
+  }
+
+  @Override
+  public void initialize() {
+    loggerFactory = StaticLoggerBinder.getSingleton().getLoggerFactory();
+    markerFactory = new BasicMarkerFactory();
+    contextAdapter = StaticMDCBinder.getSingleton().getMock();
+  }
+
+}
