+++ src/test/java/org/slf4j/simple/StaticMDCBinderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.slf4j.impl;
+package org.slf4j.simple;
 
 import org.junit.jupiter.api.Test;
